fix(nsproxy): proper NSProxy handling
