+++ TestFixtures/Interfaces/TNSClassWithPlaceholder.h
@@ -2,6 +2,4 @@
 
 @interface TNSClassWithPlaceholder : NSObject
 
-@property(readonly) TNSClassWithPlaceholder* testEmbeddedClass;
-
 @end
