+++ TestFixtures/Interfaces/TNSClassWithPlaceholder.m
@@ -25,10 +25,6 @@
     return (id)[[TNSClassWithPlaceholderReal alloc] init];
 }
 
-- (TNSClassWithPlaceholder*)testEmbeddedClass {
-    return (id)[[TNSClassWithPlaceholderReal alloc] init];
-}
-
 #pragma clang diagnostic pop
 
 - (instancetype)retain {
@@ -53,10 +49,6 @@
     }
 
     return [super alloc];
-}
-
-- (TNSClassWithPlaceholder*)testEmbeddedClass {
-    return [TNSClassWithPlaceholderPlaceholder alloc];
 }
 
 @end
