+++ TestRunner/app/tests/Infrastructure/Jasmine/jasmine-2.0.1/jasmine.js
@@ -102,7 +102,7 @@ getJasmineRequireObj().base = (function (jasmineGlobal) {
 
     j$.MAX_PRETTY_PRINT_DEPTH = 40;
     j$.MAX_PRETTY_PRINT_ARRAY_LENGTH = 100;
-    j$.DEFAULT_TIMEOUT_INTERVAL = 10000000;
+    j$.DEFAULT_TIMEOUT_INTERVAL = 5000;
 
     j$.getGlobal = function() {
       return jasmineGlobal;
